Fixed models generation.
