+++ lib/src/data/service/local_storage_service.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:path_provider/path_provider.dart' as path_provider;
+import 'package:tasker/src/data/models/priority_level.dart';
+import 'package:tasker/src/data/models/subtask.dart';
 import 'package:tasker/src/data/models/task.dart';
 
 class LocalStorageService {
   static Future<void> initLocalStorage() async {
     try {
       await Hive.initFlutter();
-      await Hive.openBox<bool>('theme');
       final appDocumentDir =
           await path_provider.getApplicationDocumentsDirectory();
       Hive.init(appDocumentDir.path);
@@ -15,6 +16,7 @@ class LocalStorageService {
       Hive.registerAdapter(SubtaskAdapter());
       Hive.registerAdapter(PriorityLevelAdapter()); // Register the enum adapter
       await Hive.openBox<Task>('tasks');
+      await Hive.openBox<bool>('theme');
     } catch (e) {
       debugPrint("ERROR ON initLocalStorage(): ${e.toString()}");
     }
